Split config.rs into config/ directory module
- Convert src/config.rs to src/config/mod.rs (396 LOC, was 570)
- Extract persistence functions (save_state, load_state, save_state_to,
  load_state_from, get_state_file_path) to src/config/persistence.rs (139 LOC)
- Move persistence-specific tests to persistence.rs
- Add tempfile as dev-dependency for persistence tests
- All 100 tests pass, zero clippy warnings
