+++ Cargo.toml
@@ -78,3 +78,6 @@ lto = true
 codegen-units = 1
 strip = true
 panic = "abort"
+
+[dev-dependencies]
+tempfile = "3.27.0"
